+++ Tabloid/Controllers/PostController.cs
@@ -4,7 +4,7 @@ using Tabloid.Repositories;
 
 namespace Tabloid.Controllers
 {
-    [Authorize]
+ 
     [Route("api/[controller]")]
     [ApiController]
     public class PostController : ControllerBase
