+++ Tabloid/client/src/components/ApplicationViews.js
@@ -4,6 +4,7 @@ import Login from "./Login";
 import Register from "./Register";
 import Hello from "./Hello";
 import TagList from "./TagList";
+import PostList from "./Posts/PostList";
 
 export default function ApplicationViews({ isLoggedIn }) {
   return (
@@ -17,6 +18,7 @@ export default function ApplicationViews({ isLoggedIn }) {
           <Route path="login" element={<Login />} />
           <Route path="register" element={<Register />} />
           <Route path="tag" element={<TagList />} />
+          <Route path="post" element={<PostList />} />
           <Route path="*" element={<p>Whoops, nothing here...</p>} />
         </Route>
       </Routes>
