+++ Tabloid/client/src/components/Header.js
@@ -31,6 +31,9 @@ export default function Header({ isLoggedIn }) {
                 <NavItem>
                   <NavLink tag={RRNavLink} to="/tag">Tag Management</NavLink>
                 </NavItem>
+                <NavItem>
+                  <NavLink tag={RRNavLink} to="/post">Post Management</NavLink>
+                </NavItem>
               </div>
             }
           </Nav>
